pyproject.toml
[tool.poetry]
name = "t1me.it"
version = "0.1.0"
description = "Make thumbnaiils for exact video timecodes"
authors = ["Egor Gusev <guy.do.or.die@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.116.1"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
playwright = "^1.54.0"
pillow = "^11.3.0"
redis = "^6.3.0"
python-multipart = "^0.0.20"
aiofiles = "^24.1.0"
pydantic = "^2.11.7"
httpx = "^0.28.1"
python-dotenv = "^1.1.1"
beautifulsoup4 = "^4.13.4"
shortuuid = "^1.0.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
